Moved tactics to the library
See https://github.com/advancedresearch/hooo/issues/12
